Create a threadRunner
